requirements.txt
django
djangorestframework
djangorestframework-jwt
django-cors-headers
psycopg2-binary
django-rest-recaptcha
